requirements.txt
asgiref==3.7.0
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==41.0.1
defusedxml==0.7.1
distlib==0.3.6
Django==4.2.1
django-allauth==0.54.0
django-cors-headers==4.2.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
filelock==3.12.0
google-auth==2.19.1
google-auth-oauthlib==1.0.0
idna==3.4
mysqlclient==2.2.0
oauthlib==3.2.2
Pillow==10.0.0
platformdirs==3.5.1
psycopg2==2.9.7
psycopg2-binary==2.9.7
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.18.0
pygame==2.4.0
PyJWT==2.7.0
pyotp==2.9.0
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
tzdata==2023.3
urllib3==1.26.16
virtualenv==20.23.0